Organize cards logic to start modularizing...
