+++ app/scripts/components/common/info-card-horizontal/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export function InfoCardHorizontal () {
+  return (
+    <>
+    </>
+  );
+}
