+++ app/scripts/components/sandbox/index.js
@@ -19,7 +19,8 @@ import SandboxMDXEditor from './mdx-editor';
 import SandboxTable from './table';
 import SandboxLayerInfo from './layer-info';
 import { resourceNotFound } from '$components/uhoh';
-import { Card, CardList } from '$components/common/card';
+import { Card } from '$components/common/card';
+import { CardList } from '$components/common/card/styles';
 import { Fold, FoldHeader, FoldTitle } from '$components/common/fold';
 import PageHero from '$components/common/page-hero';
 import { LayoutProps } from '$components/common/layout-root';
