+++ app/scripts/components/stories/hub/index.tsx
@@ -26,12 +26,8 @@ import {
   FoldHeadline,
   FoldTitle
 } from '$components/common/fold';
-import {
-  Card,
-  CardList,
-  CardMeta,
-  CardTopicsList
-} from '$components/common/card';
+import { Card } from '$components/common/card';
+import { CardList, CardMeta, CardTopicsList } from '$components/common/card/styles';
 import EmptyHub from '$components/common/empty-hub';
 import { PageMainContent } from '$styles/page';
 import { STORIES_PATH, getStoryPath } from '$utils/routes';
